requirements.txt
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3
openai==1.12.0
python-dotenv==1.0.1
umap-learn==0.5.5
matplotlib==3.8.2
